Replace use of org.khronos.webgl imports with own definitions
