+++ library/process/src/jsMain/kotlin/io/matthewnelson/kmp/process/internal/JsPlatform.kt
@@ -20,10 +20,6 @@ package io.matthewnelson.kmp.process.internal
 import io.matthewnelson.kmp.file.*
 import io.matthewnelson.kmp.process.InternalProcessApi
 import io.matthewnelson.kmp.process.ReadBuffer
-import org.khronos.webgl.ArrayBufferView
-import kotlin.contracts.ExperimentalContracts
-import kotlin.contracts.InvocationKind
-import kotlin.contracts.contract
 
 internal actual val STDIO_NULL: File by lazy {
     val isWindows = try {
@@ -39,41 +35,6 @@ internal actual val IsMobile: Boolean get() = try {
     os_platform() == "android"
 } catch (_: Throwable) {
     false
-}
-
-@Suppress("NOTHING_TO_INLINE")
-@OptIn(ExperimentalContracts::class)
-// @Throws(IllegalArgumentException::class, IndexOutOfBoundsException::class)
-internal inline fun <T: ArrayBufferView> ByteArray.toJsArray(
-    offset: Int = 0,
-    len: Int = size - offset,
-    checkBounds: Boolean = false,
-    factory: (size: Int) -> T,
-): T {
-    contract {
-        callsInPlace(factory, InvocationKind.AT_MOST_ONCE)
-    }
-
-    if (checkBounds) checkBounds(offset, len)
-    val array = factory(len)
-    val dArray = array.asDynamic()
-
-    var aI = 0
-    for (i in offset until offset + len) {
-        dArray[aI++] = this[i]
-    }
-
-    return array
-}
-
-@Suppress("NOTHING_TO_INLINE")
-internal inline fun ArrayBufferView.fill() {
-    val len = byteLength
-    if (len == 0) return
-    val a = asDynamic()
-    for (i in 0 until len) {
-        a[i] = 0
-    }
 }
 
 /** @suppress */
